Delete meow_gif.py
This commit involves removing hard coded links stored in Python lists that were previously being used by the bot. Instead, the Giphy API is now being used to fetch the links directly from https://giphy.com/.

Storing gif links in a list can be cumbersome and inflexible, especially when you need to add or remove gifs from the list. Moreover, it can be difficult to ensure that the list is always up-to-date with the latest gifs available on the internet. By using the Giphy API, the bot can now fetch the links in real-time, ensuring that it always provides the latest and greatest selection of gifs.

This change provides a more efficient and reliable way of managing the bot's gif resources. It also helps to future-proof the bot's functionality, as it can easily adapt to any changes or updates made to the Giphy API. Overall, this is a significant improvement over the previous method of storing links in a list.